+++ snooplyze/persistence/__init__.py
@@ -3,7 +3,8 @@ from .base_persistence_engine import PersistenceEngine, PersistenceEngineType
 from .duckdb_persistence_engine import DuckDbPersistenceEngine
 from .postgresql_persistence_engine import PostgreSQLPersistenceEngine
 from .flatfile_persistence_engine import FlatFilePersistenceEngine
+from .mysql_persistence_engine import MySQLPersistenceEngine
 
 
 __all__ = ["PersistenceEngine", "PersistenceEngineType",
-            "DuckDbPersistenceEngine", "PostgreSQLPersistenceEngine", "FlatFilePersistenceEngine"]
+            "DuckDbPersistenceEngine", "PostgreSQLPersistenceEngine", "FlatFilePersistenceEngine", "MySQLPersistenceEngine"]