+++ snooplyze/persistence/mysql/scripts/structure.sql
@@ -1,0 +1,7 @@
+CREATE TABLE content
+(
+    name VARCHAR(100),
+    timeadded DATETIME,
+    hash VARCHAR(100),
+    content TEXT
+)